Remove abstraction of ExternalVideoTrackSource
Remove the internal abstraction of the ExternalVideoTrackSource class,
which used to exhibit a clean C++ interface but has no reason to be now
that only a pure C interop API is exported. This simplifies the internal
implementation by removing a level of indirection.
